[fix] 팔로잉 팔로워 수 바꿔 나온 사항 수정

+++ ItNovation/src/main/java/com/ItsTime/ItNovation/domain/bestUser/TopUserResponseDto.java
@@ -17,7 +17,7 @@ public class TopUserResponseDto {
     private String nickName;
     private String introduction;
     private Grade grade;
-    private int followers;
+    private Long followers;
     private Long followings;
     private Boolean isLoginedUserFollowsNowUser;
 
@@ -27,7 +27,7 @@ public class TopUserResponseDto {
 
     @Builder
     public TopUserResponseDto(Long userId, String profileImg, String nickName, String introduction,
-                              Grade grade, int followers, Long followings, List<TopUserReviewDto> reviews, Boolean isLoginedUserFollowsNowUser) {
+                              Grade grade, Long followers, Long followings, List<TopUserReviewDto> reviews, Boolean isLoginedUserFollowsNowUser) {
         this.userId = userId;
         this.profileImg = profileImg;
         this.nickName = nickName;
